backend/postgresql_cleanup/requirements.txt
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0 